added wehub corpus api

+++ src/main/java/com/metaverse/workflow/nontraining/dto/PhysicalFinancialDto.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class PhysicalFinancialDto {
     private Long nonTrainingAchievementId;
     private Long nonTrainingActivityId;
-    private Double physicalTarget;
+    private Long physicalTarget;
     private String physicalTargetAchievement;
     private Double financialTarget;
     private Double financialTargetAchievement;
